move empty to whyis package
